chore: improved types
