+++ ts/features/idpay/wallet/components/__tests__/IdPayCard.test.tsx
@@ -1,0 +1,19 @@
+import { render } from "@testing-library/react-native";
+import * as React from "react";
+import { IdPayCard } from "../IdPayCard";
+
+describe("IdPayWalletCard", () => {
+  it("should match the snapshot", () => {
+    const component = render(
+      <IdPayCard
+        name="18 app"
+        amount={9999}
+        avatarSource={{
+          uri: "https://vtlogo.com/wp-content/uploads/2021/08/18app-vector-logo.png"
+        }}
+        expireDate={new Date(2023, 11, 2)}
+      />
+    );
+    expect(component).toMatchSnapshot();
+  });
+});
