+++ ts/features/newWallet/types/index.ts
@@ -34,20 +34,22 @@ export type WalletCardPayment = {
   type: "payment";
 } & PaymentWalletCardProps;
 
+// Base WalletCard type, which includes all card types
 export type WalletCard = WalletCardBase &
   (WalletCardBonus | WalletCardCgn | WalletCardPayment);
+
 // Used to map the card to the specific component that will render the card.
 export type WalletCardType = WalletCard["type"];
 
 /**
  * Wallet card component mapper which translates a WalletCardType to a
  * component to be rendered inside the wallet.
- * Component MUST be a WalletCardBaseComponent, which can be wrapped
- * using {@see withWalletCardBaseComponent}
+ * Component MUST be a WalletCardBaseComponent, which can be created
+ * using {@see withWalletCardBaseComponent} HOC
  */
 export const walletCardComponentMapper: Record<
   WalletCardType,
-  WalletCardBaseComponent<any> | undefined
+  WalletCardBaseComponent<any, any> | undefined
 > = {
   cgn: undefined, // TODO add CGN component type
   idPay: IdPayWalletCard,
